+++ Pipfile
@@ -1,0 +1,20 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+
+[packages]
+six = "*"
+pymongo = "*"
+alembic = "*"
+mongoengine = "*"
+Flask = "*"
+SQLAlchemy = "*"
+CouchDB = ">=1.0"
+redis = "==2.10.6"
+flask-track-usage = {editable = true,path = "."}
+
+[requires]
+python_version = "3.6"
